unix,win: unify timers implementation
PR-URL: https://github.com/libuv/libuv/pull/1058
