+++ Makefile.am
@@ -30,6 +30,7 @@ libuv_la_SOURCES = src/fs-poll.c \
                    src/loop-watcher.c \
                    src/queue.h \
                    src/threadpool.c \
+                   src/timer.c \
                    src/uv-common.c \
                    src/uv-common.h \
                    src/version.c
@@ -69,7 +70,6 @@ libuv_la_SOURCES += src/win/async.c \
                     src/win/stream-inl.h \
                     src/win/tcp.c \
                     src/win/thread.c \
-                    src/win/timer.c \
                     src/win/tty.c \
                     src/win/udp.c \
                     src/win/util.c \
@@ -99,7 +99,6 @@ libuv_la_SOURCES += src/unix/async.c \
                    src/unix/stream.c \
                    src/unix/tcp.c \
                    src/unix/thread.c \
-                   src/unix/timer.c \
                    src/unix/tty.c \
                    src/unix/udp.c
