+++ src/uv-common.h
@@ -221,4 +221,9 @@ void uv__idle_close(uv_idle_t* handle);
 void uv__prepare_close(uv_prepare_t* handle);
 void uv__check_close(uv_check_t* handle);
 
+/* Timer prototypes */
+void uv__run_timers(uv_loop_t* loop);
+int uv__next_timeout(const uv_loop_t* loop);
+void uv__timer_close(uv_timer_t* handle);
+
 #endif /* UV_COMMON_H_ */
