+++ uv.gyp
@@ -77,6 +77,7 @@
         'src/loop-watcher.c',
         'src/queue.h',
         'src/threadpool.c',
+        'src/timer.c',
         'src/uv-common.c',
         'src/uv-common.h',
         'src/version.c'
@@ -113,7 +114,6 @@
             'src/win/stream-inl.h',
             'src/win/tcp.c',
             'src/win/tty.c',
-            'src/win/timer.c',
             'src/win/udp.c',
             'src/win/util.c',
             'src/win/winapi.c',
@@ -157,7 +157,6 @@
             'src/unix/stream.c',
             'src/unix/tcp.c',
             'src/unix/thread.c',
-            'src/unix/timer.c',
             'src/unix/tty.c',
             'src/unix/udp.c',
           ],
